Move ProjectRow to where both pages can see it
The composer needs the same project rows the sidebar card renders, and `vmux_chat` does not
depend on `vmux_layout` — nor should it, the dependency runs the other way through
`vmux_setting`. `ProjectRow` is a wire type either page may carry, so it belongs in
`vmux_core::event` beside the other shapes that cross that boundary.
